+++ src/mastodon_api/oauth/utils.rs
@@ -3,39 +3,18 @@ use base64;
 use mitra_utils::random::generate_random_sequence;
 
 pub fn render_authorization_page() -> String {
-    let page = r#"<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>Authorization</title>
-    <style nonce="oauth-authorization">
-        html, body { height: 100%; }
-        form {
-            display: flex;
-            flex-direction: column;
-            gap: 5px;
-            margin: auto;
-            max-width: 100%;
-            position: relative;
-            top: 40%;
-            width: 200px;
-        }
-    </style>
-</head>
-<body>
+    format!(include_str!("template.html"), content=r#"
     <form method="post">
         <input type="text" name="username" placeholder="Username">
         <br>
         <input type="password" name="password" placeholder="Password">
         <br>
         <button type="submit">Submit</button>
-    </form>
-</body>
-</html>
-"#.to_string();
-    page
+    </form>"#)
+}
+
+pub fn render_authorization_code_page(code: String) -> String {
+    format!(include_str!("template.html"), content=code)
 }
 
 const ACCESS_TOKEN_SIZE: usize = 20;